Return full API responses from NL.
