+++ language/google/cloud/language/sentiment.py
@@ -1,4 +1,4 @@
-# Copyright 2016 Google Inc.
+# Copyright 2017 Google Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -16,41 +16,6 @@
 
 Sentiment is the response to an ``analyzeSentiment`` request.
 """
-
-from google.cloud.language.syntax import Sentence
-
-
-class SentimentResponse(object):
-    """A representation of a response to an ``analyzeSentiment`` request
-    to the Google Natural Language API.
-
-    :type sentiment: :class:`~.language.sentiment.Sentiment`
-    :param sentiment: A Sentiment object.
-
-    :type language: str
-    :param: language: The language used for analyzing sentiment.
-
-    :type sentences: list
-    :param sentences: A list of :class:`~.language.syntax.Sentence` objects.
-    """
-    def __init__(self, sentiment, language, sentences):
-        self.sentiment = sentiment
-        self.language = language
-        self.sentences = sentences
-
-    @classmethod
-    def from_api_repr(cls, payload):
-        """Return an sentiment response from a JSON representation.
-
-        :type payload: dict
-        :param payload: A dictionary representing the response.
-        """
-        return cls(
-            language=payload.get('language', None),
-            sentences=[Sentence.from_api_repr(i) for i
-                       in payload.get('sentences', ())],
-            sentiment=Sentiment.from_api_repr(payload['documentSentiment']),
-        )
 
 
 class Sentiment(object):
