+++ language/google/cloud/language/syntax.py
@@ -1,4 +1,4 @@
-# Copyright 2016 Google Inc.
+# Copyright 2017 Google Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -19,6 +19,7 @@ breaks a document down into tokens and sentences.
 """
 
 from google.cloud.language.sentiment import Sentiment
+from google.cloud.language.sentence import Sentence
 
 
 class PartOfSpeech(object):
@@ -101,37 +102,6 @@ class PartOfSpeech(object):
         return cls._REVERSE_MAP[tag]
 
 
-class SyntaxResponse(object):
-    """A representation of a response to an ``analyzeSyntax`` request
-    to the Google Natural Language API.
-
-    :type tokens: list
-    :param tokens: A list of :class:`~.language.syntax.Token` objects.
-
-    :type language: str
-    :param: language: The language used for analyzing sentiment.
-
-    :type sentences: list
-    :param sentences: A list of :class:`~.language.syntax.Sentence` objects.
-    """
-    def __init__(self, tokens, language, sentences):
-        self.tokens = tokens
-        self.language = language
-        self.sentences = sentences
-
-    @classmethod
-    def from_api_repr(cls, payload):
-        """Return an syntax response from a JSON representation.
-
-        :type payload: dict
-        :param payload: A dictionary representing the response.
-        """
-        return cls(
-            language=payload.get('language', None),
-            sentences=[Sentence(i) for i in payload.get('sentences', ())],
-            tokens=[Token(i) for i in payload.get('tokens', ())]
-        )
-
 class Token(object):
     """A Google Cloud Natural Language API token object.
 
@@ -199,53 +169,3 @@ class Token(object):
         lemma = payload['lemma']
         return cls(text_content, text_begin, part_of_speech,
                    edge_index, edge_label, lemma)
-
-
-class Sentence(object):
-    """A Google Cloud Natural Language API sentence object.
-
-    .. _Sentence message: https://cloud.google.com/natural-language/reference\
-                          /rest/v1/documents/annotateText#Sentence
-
-    See `Sentence message`_.
-
-    :type content: str
-    :param content: The text that the sentence is composed of.
-
-    :type begin: int
-    :param begin: The beginning offset of the sentence in the original
-                  document according to the encoding type specified
-                  in the API request.
-
-    :type sentiment: :class:`~google.cloud.language.sentiment.Sentiment`
-    :param sentiment:
-        (Optional) For calls to
-        :meth:`~google.cloud.language.document.Document.annotate_text` where
-        ``include_sentiment`` is set to true, this field will contain the
-        sentiment for the sentence.
-    """
-
-    def __init__(self, content, begin, sentiment=None):
-        self.content = content
-        self.begin = begin
-        self.sentiment = sentiment
-
-    @classmethod
-    def from_api_repr(cls, payload):
-        """Convert a sentence from the JSON API into a :class:`Sentiment`.
-
-        :param payload: dict
-        :type payload: The value from the backend.
-
-        :rtype: :class:`Sentence`
-        :returns: The sentence parsed from the API representation.
-        """
-        text_span = payload['text']
-
-        try:
-            sentiment = Sentiment.from_api_repr(payload['sentiment'])
-        except KeyError:
-            sentiment = None
-
-        return cls(text_span['content'], text_span['beginOffset'],
-                   sentiment=sentiment)
