+++ language/unit_tests/test_syntax.py
@@ -95,62 +95,3 @@ class TestToken(unittest.TestCase):
         self.assertEqual(token.edge_index, edge_index)
         self.assertEqual(token.edge_label, edge_label)
         self.assertEqual(token.lemma, lemma)
-
-
-class TestSentence(unittest.TestCase):
-
-    @staticmethod
-    def _get_target_class():
-        from google.cloud.language.syntax import Sentence
-
-        return Sentence
-
-    def _make_one(self, *args, **kw):
-        return self._get_target_class()(*args, **kw)
-
-    def test_constructor(self):
-        content = "All the king's horses."
-        begin = 11
-        sentence = self._make_one(content, begin)
-        self.assertEqual(sentence.content, content)
-        self.assertEqual(sentence.begin, begin)
-
-    def test_from_api_repr(self):
-        klass = self._get_target_class()
-        content = 'All the pretty horses.'
-        begin = -1
-        payload = {
-            'text': {
-                'content': content,
-                'beginOffset': begin,
-            },
-        }
-        sentence = klass.from_api_repr(payload)
-        self.assertEqual(sentence.content, content)
-        self.assertEqual(sentence.begin, begin)
-        self.assertEqual(sentence.sentiment, None)
-
-    def test_from_api_repr_with_sentiment(self):
-        from google.cloud.language.sentiment import Sentiment
-
-        klass = self._get_target_class()
-        content = 'All the pretty horses.'
-        begin = -1
-        score = 0.5
-        magnitude = 0.5
-        payload = {
-            'text': {
-                'content': content,
-                'beginOffset': begin,
-            },
-            'sentiment': {
-                'score': score,
-                'magnitude': magnitude,
-            }
-        }
-        sentence = klass.from_api_repr(payload)
-        self.assertEqual(sentence.content, content)
-        self.assertEqual(sentence.begin, begin)
-        self.assertIsInstance(sentence.sentiment, Sentiment)
-        self.assertEqual(sentence.sentiment.score, score)
-        self.assertEqual(sentence.sentiment.magnitude, magnitude)
